changed: put the voigt tensor helper functions in a shared source
they were duplicated in 3 source files
